data persistence added

+++ BusShuttleProject/BusShuttle/FileSaver.cs
@@ -6,8 +6,11 @@ public class FileSaver
     public FileSaver(string fileName)
     {
         this.fileName = fileName;
-        File.Create(this.fileName).Close();
+        if(!File.Exists(this.fileName)) {
+            File.Create(this.fileName).Close();
+        }
     }
+
     public void AppendLine(string line)
     {
         File.AppendAllText(this.fileName, line + Environment.NewLine);
